Fixed bugs and additional functions
I fixed the bug when script was searching not for many masters but form many changes to master.
I also added test mode.
Also added some obvious exceptions (when master is empty, when master is space string or when name is reserved in Update,esm for cc content). There is a space to add some more exceptions (fe. in configured array).
There is also added some simple output for analyzing.